Properly modularize Minitest plugins
